Refactor angular package class
