+++ index.ts
@@ -1,95 +1,6 @@
 import { posix as path } from 'path';
 
-import { AngularPackageConfig, AngularSubPackageConfig } from './src/config.interface';
-import { readFile } from './src/utilities/read-file';
-
-// TODO: Find a better name?
-export interface AngularPackageConfigInternal {
-	entryFile: string;
-	outDir: string;
-	packageName: string;
-	fileName: string;
-}
-
-/**
- * Angular Package
- */
-export class AngularPackage {
-
-	/**
-	 * Angular Package project directory
-	 */
-	private cwd: string;
-
-	/**
-	 * Primary entry point
-	 */
-	private primaryEntry: AngularPackageConfigInternal;
-
-	/**
-	 * List of secondary entry points
-	 */
-	private secondaryEntries: Array<AngularPackageConfigInternal>;
-
-	/**
-	 * Custom TypeScript compiler options
-	 */
-	private typescriptCompilerOptions: { [ option: string ]: any };
-
-	/**
-	 * Custom Angular compiler options
-	 */
-	private angularCompilerOptions: { [ option: string ]: any };
-
-	public async withConfig( absoluteAngularPackageJsonPath: string ): Promise<void> {
-
-		this.cwd = path.dirname( absoluteAngularPackageJsonPath );
-
-		// Read files
-		const angularPackageJson: AngularPackageConfig = await readFile( absoluteAngularPackageJsonPath );
-		const absolutePackageJsonPath: string = path.join( path.dirname( absoluteAngularPackageJsonPath ), 'package.json' );
-		const packageJson: any = await readFile( absolutePackageJsonPath );
-
-		// Get primary entry information
-		this.primaryEntry = {
-			entryFile: path.join( path.dirname( absoluteAngularPackageJsonPath ), angularPackageJson.entryFile ),
-			fileName: packageJson.name.split( '/' ).pop(),
-			outDir: path.join( path.dirname( absoluteAngularPackageJsonPath ), angularPackageJson.outDir ),
-			packageName: packageJson.name
-		};
-
-		// Get compiler options
-		this.typescriptCompilerOptions = angularPackageJson.typescriptCompilerOptions || {};
-		this.angularCompilerOptions = angularPackageJson.angularCompilerOptions || {};
-
-		// Get secondary entry information
-		this.secondaryEntries = ( angularPackageJson.secondaryEntries || [] )
-			.map( ( secondaryEntry: AngularSubPackageConfig ): AngularPackageConfigInternal => {
-
-				const absoluteSecondaryEntryFilePath: string =
-					path.join( path.dirname( absoluteAngularPackageJsonPath ), secondaryEntry.entryFile );
-				const secondaryEntryFolder: string =
-					path.relative( path.dirname( this.primaryEntry.entryFile ), path.dirname( absoluteSecondaryEntryFilePath ) );
-				const secondaryPackageName: string =
-					path.join( this.primaryEntry.packageName, secondaryEntryFolder );
-
-				return {
-					entryFile: absoluteSecondaryEntryFilePath,
-					fileName: secondaryPackageName.split( '/' ).pop(),
-					outDir: path.join( this.primaryEntry.outDir, secondaryPackageName ),
-					packageName: secondaryPackageName
-				};
-
-			} );
-
-		// TODO: Dependencies
-
-		// TODO: Angular Package JSON Schema Validation
-		// TODO: Error Handling
-
-	}
-
-}
+import { AngularPackage } from './src/angular-package';
 
 /**
  * Run Angular Package Builder
